+++ config/initializers/rack_attack.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
-require "decidim-app/rack_attack"
-require "decidim-app/rack_attack/throttling"
-require "decidim-app/rack_attack/fail2ban"
+require "decidim_app/rack_attack"
+require "decidim_app/rack_attack/throttling"
+require "decidim_app/rack_attack/fail2ban"
 
 # Enabled by default in production
 # Can be deactivated with 'ENABLE_RACK_ATTACK=0'
