+++ lib/decidim_app/config.rb
