+++ lib/decidim_app/k8s/configuration.rb
